devel/squashfs-tools: Upgrade to 4.5.1
Besides a security fix for CVE-2021-41072 this upgrade brings successful
compilation back to squashfs-tools, after the recent GCC upgrade broke
it (see https://github.com/plougher/squashfs-tools/pull/83).

The recipe didn't work as-is with the new version, though, because the
creation of the man page directory failed (as the Makefile tried to
create it in the build script's directory):

mkdir: cannot create directory '/usr/local': Read-only file system
../generate-manpages/install-manpages.sh: Creating manpage install directory failed.  Aborting

Thus, the recipe was refactored and some cruft was removed that is not
necessary anymore.