Add Linux rootfs classes
Add convenience methods for creating a Linux root fs system image. The
base rootfs class adds methods for setting infos like the hostname or
the timezone. It's also possible to create additional users or
create/special permissions on device nodes/files. For doing so, the
brtools are used and we look for special files in the .bob directories
of all dependencies. For additional users the naming convention is
.bob/PKG_NAME.user-table. Similar for device descriptions the naming is
.bob/PKG_NAME.device-table.

There is also a central rootfsPrepare method which does this
automatically by using CONFIG_ environment variables. Right now it
supports:

 CONFIG_HOSTNAME:     the target hostname; defaults to "demo"
 CONFIG_TZ:           the target timezone; defaults to "Europe/Berlin"
 CONFIG_ROOT_PW:      the target root password; defaults to "root"
 CONFIG_ROOT_PW_SALT: the target root password salt; defaults to be
                      empty and therefore is auto generated

The actual image is created by one of the rootfs::* classes. It uses a
fake root environment to be able to set e.g. the root users id. This
initial commit has support for ext4, tar and cpio images. The simplest
usage would look something like this:

  inherit: [rootfs::ext4]

  ...

  buildScript: |
    rootfsCopy "$PWD/install" "${@:2}"

    rootfsPrepare "$PWD/install" "$PWD" "${@:2}"

    rootfsExt4CreateImage "$PWD/install" "demo"

This will produce an ext4 image named $PWD/demo.img.

Please note, that you will need a base root file directory structure
which contains e.g. an /etc/passwd file or the timezone files. This is
not provided by these classes and has to come from somewhere else.