net::openssh: add libxcrypt dependency
Since glibc 2.39 it doesn't come with a libcrypt anymore. This means
sshd had no support for password decryption methods beside md5 and
openssl des.

This broke sshd login support e.g. with the rootfs support class,
because this uses sha-512 encryption of the user passwords.

Fix this by adding libxcrypt as dependency which provides this
functionality to openssh again.