utils::uboot-tools: change build mode
Don't fake a config anymore, but use the default config provided by
uboot. This is the sandbox version which should be suitable to build the
host tools.

Using a default config enables FIT image support and especially FIT
signature support within the uboot tools (especially for mkimage).

We manually remove the mkeficapsule tool, because this has dependencies
which are problematic right now.