pyyaml/pyyaml: fix cross compilation
It turns out that using pip does not bode well with cross compilation.
The reason is that pip will clean the PYTHONPATH which will prevent the
respective sysconfig module to be found. A typical error message during
pip install will look like the following:

  ModuleNotFoundError: No module named '_sysconfigdata__linux_aarch64-linux-gnu'

By using the setuptools method, everything works as expected. While at
it, also add libyaml so that the module benefits of the C parser
speedup.