Add linux build classes
Building linux is a common task required in most projects. Right now,
the kernel::linux recipe can be used for that. However, this builds a
fixed version from kernel.org, which will be not correct in a lot of
cases. Especially if one builds a project for a specific arm board, the
upstream kernel is most of the time not usable. Vendors of those boards
tend to provided there own BSP's, including a modified linux version. On
top of that, sometimes it is also required to make modifications which a
project specific.

In those cases I don't want to copy the linux.yaml all the time and make
local modifications. Instead provide a set of base class helper
functions, which make the task of compiling a linux kernel (and all
dependent task, like dtb handling) only a matter of adding a couple of
lines to the custom linux kernel recipe.

The new classes also add support by providing functions for a kmod
distribution. This allows building external kernel mode drivers against
the kernel in use.