Add a generic virtual::kernel::linux package
Make it possible to override the linux kernel used in a project. Just
set CONFIG_SELECT_LINUX to your linux package. This recipe needs to
provide targets as seen in the virtual::kernel::linux package. For a
template on how this can look like, see the documentation in
classes/linux.yaml or recipes/kernel/linux.yaml.

Any recipe that has a dependency on the linux kernel should depend on
one of the targets defined here. This is especially true for any
external linux kernel mode driver.