bsp::grub2: Call updateConfigAllRecursive manually
Required to make the checkout deterministic in all cases.

Consider the following. If the bob workspace is clean (i.e. fresh
checkout), the following sequence happens with regards to bsp::grub2:

  checkout_scm
  update_config
  bootstrap # checks out gnulib

This leads to an as-is gnulib/build-aux/config.guess because it is
checked out after update_config is run. If on the other hand
dev/src/bsp/grub2 is not removed completely but bob still updates the
workspace (e.g. after a git clean -dxf in the toplevel recipes
directory), updateConfigAllRecursive will update the config files in the
existing gnulib directory as well, resulting in a diff to the previous
variant and therefore to an indeterministic SCM checkout.

To solve this, set APPLY_UPDATE_CONFIG to "no" and call
updateConfigAllRecursive manually after bootstrap.