core::busybox: make config file source more flexible
Add more options on how the config is gathered. Bring this in line with
e.g. the linux kernel, where you could specify a custom package with a
defconfig or a busybox defconfig or the standard defconfig.

Note, the creation of debugging symbols now depends on the
BASEMENT_DEBUG variable. If set, the CONFIG_DEBUG option is set
automatically.

Also only regenerate the .config if the source config has changed.