classes/linux-kmod-dist: fix host tools rebuild
Newer scripts check for CC != HOSTCC to detect cross compilation. If
detected they rebuild the host tools for the target, because the result
of install-extmod-build is likely to be executed on the target (e.g.
they create .deb packages). This is of course not the case for us, so
fake it by unsetting them. This is not a problem because normally
nothing for the target gets compiled.