graphics::fonts::font-util: fix fontrootdir value
Some packages will use the fontutil.pc file to get the font root
directory. As pc files are relocated in Bob, the value must be hard
coded path.

Interestingly, pkg-config has a heuristic that detects if variables
start with the same string as the "prefix" variable. But we only want
the include and library directories to be relocated. To work around this
problem, use a dummy prefix. It will be overwritten anyway by the actual
pc file location.

On the other hand, the "mapdir" path must be relocated. It is used to
find the mapping data at runtime when the tools are invoked. This is at
cross-compile time of other packages.