libs::mesa3d: fix dridriverdir value
Some packages will use the dri.pc file to get the DRI drivers directory.
As pc files are relocated in Bob, the value must be hard coded path.

Interestingly, pkg-config has a heuristic that detects if variables
start with the same string as the "prefix" variable. But we only want
the include and library directories to be relocated. To work around this
problem, use a dummy prefix. It will be overwritten anyway by the actual
pc file location.