graphics::fonts::font-util: make a pure binary package
The font-utils package produces only two executables with some
development meta data. It's used by other packages to process font data
at compile time. In particular, it is *not* a library. Hence, it makes
no sense to have a -dev and -tgt multiPackage.

Because some packages require pkg-config to find the font-util binaries,
add a class that adds the tool to the pkg-config search path.